[Architecture/UI/Hazards] Inline cleanups + RoamingMonster split | Playtest: monsters, UI HUD, day/night
UI:
- Drop redundant carriedText. Inventory HUD slots already show item name and
  value per-slot, so the separate "Carrying: ..." line was duplicate noise.
- Replace FriendSlopUI.Hud's per-frame FindFirstObjectByType<DayNightCycle>
  fallback with a DayNightCycleRegistry lookup. DayNightCycle now self-
  registers in OnNetworkSpawn / OnNetworkDespawn, mirroring
  RoundManagerRegistry / LocalPlayerRegistry. Survives planet scene unloads
  cleanly.

Hazards:
- Promote RoamingMonster._isDead from a per-client bool driven by Die/Revive
  ClientRpcs to a NetworkVariable<bool>. Late-joiners now get the correct
  value via OnValueChanged instead of a missed RPC. The two ClientRpcs are
  now redundant and removed.
- Scale ComputeBodyVisibility linecast count by distance: 4 samples at close
  range, 2 at medium, 1 (torso only) at long range. With 4 players and 2
  monsters this drops the worst-case 32 linecasts/frame to 8 when no players
  are close, while keeping precision where stealth gameplay actually matters.

Refactor:
- Split RoamingMonster.cs (was 489-line baseline, would have hit 503 after
  the changes above) into three partials: main (state machine + lifecycle),
  .Perception.cs (FindNearestPlayer / ComputeBodyVisibility / CanDetectPlayer
  / IsOccluded), .Movement.cs (MoveToward / Pick*Target / TryAttack /
  AlignToSurface). Main file now 263 lines.

Tests:
- ArchitectureGuardrailTests baselines: drop RoamingMonster.cs (now 263
  lines, under default 400). Ratchet FriendSlopUI.cs (488 -> 477) and
  FriendSlopUI.BuildUi.cs (518 -> 517) to match the new actuals so any
  regression fails CI immediately.

Chore:
- Add /dev/ and .claude/ to .gitignore. The dev/null/ directory was a Windows
  side-effect of git LFS hooks treating /dev/null as a literal path.
