+++ Space Game/Assets/Scripts/Core/DayNightCycleRegistry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db8ade992c2142d4a15652e3624e3002
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
