+++ Space Game/Assets/Scripts/Hazards/RoamingMonster.Movement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f2cfb8af7194b15808cc2a0f8189b27
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
