+++ Space Game/Assets/Scripts/Hazards/RoamingMonster.Perception.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e9ebb91621ce405f8e1c107cc01736bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
